+++ src/core/main.py
@@ -37,6 +37,7 @@ from src.utils import settings
 from src.utils import session_handler
 from src.utils import simple_http_server
 from src.thirdparty.colorama import Fore, Back, Style, init
+from src.core.testing import smoke_test
 from src.core.requests import tor
 from src.core.requests import proxy
 from src.core.requests import headers
@@ -58,6 +59,7 @@ if settings.IS_WINDOWS:
   codecs.register(lambda name: codecs.lookup("utf-8") if name == "cp65001" else None)
   # Use Colorama to make Termcolor work on Windows too :)
   init()
+
 
 """
 Define HTTP User-Agent header.
@@ -627,6 +629,10 @@ def main(filename, url):
     #raise SystemExit()
 
 try:
+
+  if menu.options.smoke_test:
+    smoke_test()
+
   # Check if defined "--version" option.
   if menu.options.version:
     version.show_version()